DecimalField instead of FloatField,load humanize,floatformat'

+++ charcoallog/core/models.py
@@ -103,7 +103,7 @@ class ExtractManager(models.Manager):
 class Extract(models.Model):
     user_name = models.CharField('Name', max_length=30)
     date = models.DateField('Date')
-    money = models.FloatField('Money', default=00.00, null=False, blank=False)
+    money = models.DecimalField('Money', default=00.00, max_digits=10, decimal_places=2, null=False, blank=False)
     description = models.CharField('Description', max_length=70)
     category = models.CharField('Category', max_length=70)
     payment = models.CharField('Payment', max_length=70)
